Implement proper handling of task's run history
