+++ gridcapa-task-manager-app/src/main/java/com/farao_community/farao/gridcapa/task_manager/app/service/StatusHandler.java
@@ -89,5 +89,4 @@ public class StatusHandler {
         taskUpdateNotifier.notify(savedTask, true, false);
         LOGGER.info("Task status has been updated on {} to {}", task.getTimestamp(), savedTask.getStatus());
     }
-
 }
