+++ gridcapa-task-manager-app/src/test/java/com/farao_community/farao/gridcapa/task_manager/app/service/TaskServiceTest.java
@@ -1,0 +1,79 @@
+/*
+ * Copyright (c) 2024, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.farao_community.farao.gridcapa.task_manager.app.service;
+
+import com.farao_community.farao.gridcapa.task_manager.app.TaskUpdateNotifier;
+import com.farao_community.farao.gridcapa.task_manager.app.entities.FileEventType;
+import com.farao_community.farao.gridcapa.task_manager.app.entities.ProcessFile;
+import com.farao_community.farao.gridcapa.task_manager.app.entities.Task;
+import com.farao_community.farao.gridcapa.task_manager.app.repository.ProcessFileRepository;
+import com.farao_community.farao.gridcapa.task_manager.app.repository.TaskRepository;
+import com.farao_community.farao.minio_adapter.starter.MinioAdapter;
+import com.farao_community.farao.minio_adapter.starter.MinioAdapterConstants;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.actuate.autoconfigure.metrics.CompositeMeterRegistryAutoConfiguration;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.cloud.stream.function.StreamBridge;
+
+import java.time.OffsetDateTime;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+/**
+ * @author Vincent Bochet {@literal <vincent.bochet at rte-france.com>}
+ */
+@SpringBootTest
+class TaskServiceTest {
+    private static final String INPUT_FILE_GROUP_VALUE = MinioAdapterConstants.DEFAULT_GRIDCAPA_INPUT_GROUP_METADATA_VALUE;
+
+    @Autowired
+    private TaskUpdateNotifier taskUpdateNotifier;
+
+    @MockBean
+    private StreamBridge streamBridge; // Useful to avoid AMQP connection that would fail
+
+    @MockBean
+    private MinioAdapter minioAdapter;
+
+    @Autowired
+    private ProcessFileRepository processFileRepository;
+
+    @Autowired
+    private TaskRepository taskRepository;
+
+    @Autowired
+    private TaskService taskService;
+    @Autowired
+    private CompositeMeterRegistryAutoConfiguration compositeMeterRegistryAutoConfiguration;
+
+    @AfterEach
+    void cleanDatabase() {
+        taskRepository.deleteAll();
+        processFileRepository.deleteAll();
+    }
+
+    @Test
+    void addFileEventToTaskWhenManualUploadTest() {
+        Task task = new Task(OffsetDateTime.now());
+        ProcessFile processFile = new ProcessFile(
+                "path/MANUAL_UPLOAD/to/cgm-file.xml",
+                "input",
+                "CGM",
+                OffsetDateTime.parse("2021-10-11T00:00Z"),
+                OffsetDateTime.parse("2021-10-12T00:00Z"),
+                OffsetDateTime.parse("2021-10-11T10:18Z"));
+
+        taskService.addFileEventToTask(task, FileEventType.UPDATED, processFile);
+
+        assertEquals(1, task.getProcessEvents().size());
+        assertTrue(task.getProcessEvents().first().getMessage().startsWith("Manual upload of a"));
+    }
+}
